requirements.txt
streamlit
opencv-python
mediapipe-silicon
numpy
ultralytics
scipy